# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Papi Foundation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ongoing, completed, upcoming]
        - name: country
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /donations:
    get:
      operationId: listDonations
      tags: [donations]
      summary: List donations
      responses:
        "200":
          description: Donations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Donation"
    post:
      operationId: createDonation
      tags: [donations]
      summary: Create a donation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
...
  /donations/checkout:
    post:
      operationId: createDonationCheckout
      tags: [donations]
      summary: Create a Stripe Checkout session for a donation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonationCheckoutInput"
      responses:
        "200":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonationCheckoutSession"
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events and updates
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
...
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get a single event by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
...
  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
  /newsletter:
    post:
      operationId: subscribeNewsletter
      tags: [contact]
      summary: Subscribe to newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterInput"
      responses:
        "201":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterResponse"
components:
  schemas:
...
    Project:
      type: object
      required: [id, title, description, status, country, category, goalAmount, raisedAmount, imageUrl, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [ongoing, completed, upcoming]
        country:
          type: string
        category:
          type: string
        goalAmount:
          type: number
        raisedAmount:
          type: number
        progressPercent:
          type: number
        imageUrl:
          type: string
        beforeImageUrl:
          type: ["string", "null"]
        afterImageUrl:
          type: ["string", "null"]
        beneficiaries:
          type: ["integer", "null"]
        location:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        createdAt:
          type: string
    ProjectInput:
      type: object
      required: [title, description, status, country, category, goalAmount, imageUrl]
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [ongoing, completed, upcoming]
        country:
          type: string
        category:
          type: string
        goalAmount:
          type: number
        raisedAmount:
          type: number
        imageUrl:
          type: string
        beforeImageUrl:
          type: string
        afterImageUrl:
          type: string
        beneficiaries:
          type: integer
        location:
          type: string
        lat:
          type: number
        lng:
          type: number
...
    Donation:
      type: object
      required: [id, amount, currency, donorName, projectId, createdAt]
      properties:
        id:
          type: integer
        amount:
          type: number
        currency:
          type: string
        donorName:
          type: string
        isAnonymous:
          type: boolean
        projectId:
          type: ["integer", "null"]
        message:
          type: ["string", "null"]
        type:
          type: string
          enum: [one-time, monthly]
        createdAt:
          type: string
    DonationInput:
      type: object
      required: [amount, currency, donorName]
      properties:
        amount:
          type: number
        currency:
          type: string
        donorName:
          type: string
        isAnonymous:
          type: boolean
        projectId:
          type: integer
        message:
          type: string
        type:
          type: string
          enum: [one-time, monthly]
...
    DonationCheckoutInput:
      type: object
      required: [amount, currency, donorName, type]
      properties:
        amount:
          type: number
          exclusiveMinimum: 0
          maximum: 1000000
        currency:
          type: string
          enum: [EUR, USD, GBP]
        donorName:
          type: string
          minLength: 1
          maxLength: 200
        isAnonymous:
          type: boolean
        projectId:
          type: integer
        message:
          type: string
          maxLength: 1000
        type:
          type: string
          enum: [one-time, monthly]
    DonationCheckoutSession:
      type: object
      required: [url]
      properties:
        url:
          type: string
...
    Event:
      type: object
      required: [id, title, description, type, imageUrl, date, location, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        content:
          type: ["string", "null"]
        type:
          type: string
          enum: [event, update, press, video]
        imageUrl:
          type: string
        date:
          type: string
        location:
          type: string
        videoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    EventInput:
      type: object
      required: [title, description, type, imageUrl, date, location]
      properties:
        title:
          type: string
        description:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [event, update, press, video]
        imageUrl:
          type: string
        date:
          type: string
        location:
          type: string
        videoUrl:
          type: string
...
    ContactInput:
      type: object
      required: [name, email, subject, message]
      properties:
        name:
          type: string
        email:
          type: string
        subject:
          type: string
        message:
          type: string
        phone:
          type: string
    ContactResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
    NewsletterInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
        name:
          type: string
    NewsletterResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string